Added granularity and shuffling to new methods. Changed data types and added more warnings and checks.

+++ Application/Application.h
@@ -27,7 +27,7 @@ namespace NetworkLoad {
     private:
         Configuration configuration;
         std::array<char, MPI_MAX_PROCESSOR_NAME> processorName;
-        int size, rank;
+        size_t size, rank;
     };
 
 }
